+++ lib/src/features/widgets/buttons/core/utils/methods/get_fluent_button_style_method.dart
@@ -11,9 +11,9 @@ import 'package:pactus_gui_widgetbook/src/core/theme/app_theme.dart';
 /// The purpose of this function is to return a `ButtonStyle` object that controls how a button appears and behaves visually based on its state (pressed, hovered, disabled) and the provided configuration (outlined, padding size).
 ///
 /// ### Parameters:
-/// - **[context]** (`BuildContext`): The context is used to access the app's theme for applying consistent colors and styles.
-/// - **[paddingSize]** (`PaddingSizeEnum`): Specifies the amount of padding to apply horizontally inside the button. This affects the layout of the button content.
-///
+/// - [context] (`BuildContext`): The context is used to access the app's theme for applying consistent colors and styles.
+/// - [paddingSize] (`PaddingSizeEnum`): Specifies the amount of padding to apply horizontally inside the button. This affects the layout of the button content.
+/// - [borderRadius] (`double`): Controls the corner rounding of the button's border.
 /// ### Function Logic:
 /// - **Padding**: The padding is set based on the provided `paddingSize`. The horizontal padding is determined by the value of `paddingSize.horizontalSize`.
 /// - **Background Color**: The background color of the button changes based on the button state (hovered, pressed, disabled, or default). If the button is outlined, the background is transparent.
@@ -29,6 +29,7 @@ import 'package:pactus_gui_widgetbook/src/core/theme/app_theme.dart';
 ButtonStyle getFluentButtonStyleMethod({
   required BuildContext context,
   required PaddingSizeEnum paddingSize,
+  required double borderRadius,
 }) {
   final theme = AppTheme.of(context);
 
@@ -58,7 +59,7 @@ ButtonStyle getFluentButtonStyleMethod({
       return 0.0;
     }),
     shape: WidgetStateProperty.all(RoundedRectangleBorder(
-      borderRadius: BorderRadius.circular(10),
+      borderRadius: BorderRadius.circular(borderRadius),
     )),
   );
 }